requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cssselect2==0.8.0
dataclasses-json==0.6.7
ddgs==9.6.1
distro==1.9.0
dnspython==2.8.0
duckduckgo_search==8.1.1
exceptiongroup==1.3.0
filetype==1.2.0
fonttools==4.60.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.42.1
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.2
langchain-google-genai==3.0.0
langchain-mongodb==0.7.1
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==2.1.2
langgraph-checkpoint-mongodb==0.2.1
langgraph-checkpoint-postgres==3.0.0
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.37
lark==1.3.1
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.9.0
numpy==2.2.6
openai==2.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg==3.2.12
psycopg-pool==3.2.7
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pydyf==0.11.0
pymongo==4.15.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.44
sqlite-vec==0.1.6
streamlit==1.50.0
tenacity==9.1.2
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
weasyprint==66.0
webencodings==0.5.1
xxhash==3.6.0
yarl==1.22.0
zopfli==0.2.3.post1
zstandard==0.25.0
